+++ src/main/java/com/feelory/feelory_backend/words/controller/DailyWordsController.java
@@ -3,6 +3,7 @@ package com.feelory.feelory_backend.words.controller;
 import com.feelory.feelory_backend.global.api.ApiResponse;
 import com.feelory.feelory_backend.global.api.SuccessCode;
 import com.feelory.feelory_backend.global.exception.exceptions.common.InvalidDateFormatException;
+import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
 import com.feelory.feelory_backend.words.docs.DailyWordsDocs;
 import com.feelory.feelory_backend.words.model.*;
 import com.feelory.feelory_backend.words.service.DailyWordsService;
@@ -22,6 +23,7 @@ import java.time.format.DateTimeParseException;
 public class DailyWordsController {
 
     private final DailyWordsService dailyWordsService;
+    private final JwtTokenProvider jwtTokenProvider;
 
     @GetMapping("/{topicDate}")
     public ApiResponse<DailyWordDetailResponse> getDailyWord(@PathVariable String topicDate) {
@@ -45,15 +47,14 @@ public class DailyWordsController {
         return ApiResponse.success(response, SuccessCode.GET_TODAY_DAILY_WORD_SUCCESS);
     }
 
-    /*
-        TODO. [TR-YOO] Admin 검증 로직 추가 필요
-    */
     @Operation(
             summary = "오늘의 단어 추가",
             description = DailyWordsDocs.POST_DAILY_WORD_DESCRIPTION
     )
     @PostMapping("")
     public ApiResponse<DailyWordCreateResponse> postDailyWord(@Valid @RequestBody DailyWordCreateRequest request) {
+        jwtTokenProvider.checkAdmin();
+
         DailyWordCreateResponse response = dailyWordsService.registerAndUpdateDailyWord(request);
 
         boolean isAlreadyAssigned = response.getIsAlreadyAssigned();
@@ -65,22 +66,20 @@ public class DailyWordsController {
         return ApiResponse.success(response, code);
     }
 
-    /*
-        TODO. [TR-YOO] Admin 검증 로직 추가 필요
-    */
     @PatchMapping("")
     public ApiResponse<DailyWordUpdateResponse> patchDailyWord(@Valid @RequestBody DailyWordUpdateRequest request) {
+
+        jwtTokenProvider.checkAdmin();
 
         DailyWordUpdateResponse response = dailyWordsService.modifyDailyWord(request);
 
         return ApiResponse.success(response, SuccessCode.UPDATE_DAILY_WORD_SUCCESS);
     }
 
-    /*
-        TODO. [TR-YOO] Admin 검증 로직 추가 필요
-    */
     @DeleteMapping("")
     public ApiResponse<DailyWordDeleteResponse> deleteDailyWord(@Valid DailyWordDeleteRequest request) {
+
+        jwtTokenProvider.checkAdmin();
 
         DailyWordDeleteResponse response = dailyWordsService.removeDailyWord(request);
