+++ src/main/java/com/feelory/feelory_backend/writings/controller/WritingGoalsController.java
@@ -2,6 +2,7 @@ package com.feelory.feelory_backend.writings.controller;
 
 import com.feelory.feelory_backend.global.api.ApiResponse;
 import com.feelory.feelory_backend.global.api.SuccessCode;
+import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
 import com.feelory.feelory_backend.writings.model.*;
 import com.feelory.feelory_backend.writings.service.WritingGoalsService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -18,76 +19,68 @@ import org.springframework.web.bind.annotation.*;
 public class WritingGoalsController {
 
     private final WritingGoalsService writingGoalsService;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    /*
-        TODO. [TR-YOO] 로그인 유저 검증 로직 필요
-    */
     @Operation(
             summary = "글쓰기 목표 목록 조회",
             description = "글쓰기 목표 목록 조회 API"
     )
     @GetMapping("")
     public ApiResponse<WritingGoalListResponse> getWritingGoals(WritingGoalListRequest request) {
+        jwtTokenProvider.checkUserOrAdmin();
 
         WritingGoalListResponse response = writingGoalsService.getWritingGoals(request);
 
         return ApiResponse.success(response, SuccessCode.GET_WRITING_GOAL_LIST_SUCCESS);
     }
 
-    /*
-        TODO. [TR-YOO] 로그인 유저 검증 로직 필요
-    */
     @Operation(
             summary = "글쓰기 목표 상세 조회",
             description = "글쓰기 목표 상세 조회 API"
     )
     @GetMapping("/{id}")
     public ApiResponse<WritingGoalDetailResponse> getWritingGoalDetail(@PathVariable Long id) {
+        jwtTokenProvider.checkUserOrAdmin();
 
         WritingGoalDetailResponse response = writingGoalsService.getWritingGoalDetail(id);
 
         return ApiResponse.success(response, SuccessCode.GET_WRITING_GOAL_DETAIL_SUCCESS);
     }
 
-    /*
-        TODO. [TR-YOO] 로그인 유저 검증 로직 필요
-    */
     @Operation(
             summary = "글쓰기 목표 추가",
             description = "글쓰기 목표 추가 API"
     )
     @PostMapping("")
     public ApiResponse<WritingGoalCreateResponse> postWritingGoal(@Valid @RequestBody WritingGoalCreateRequest request) {
+        jwtTokenProvider.checkUserOrAdmin();
 
         WritingGoalCreateResponse response = writingGoalsService.registerWritingGoal(request);
 
         return ApiResponse.success(response, SuccessCode.REGISTER_WRITING_GOAL_SUCCESS);
     }
 
-    /*
-        TODO. [TR-YOO] 로그인 유저 검증 로직 필요
-    */
     @Operation(
             summary = "글쓰기 목표 수정",
             description = "글쓰기 목표 수정 API"
     )
     @PatchMapping("")
     public ApiResponse<WritingGoalUpdateResponse> patchWritingGoal(@Valid @RequestBody WritingGoalUpdateRequest request) {
+        jwtTokenProvider.checkUserOrAdmin();
 
         WritingGoalUpdateResponse response = writingGoalsService.modifyWritingGoal(request);
 
         return ApiResponse.success(response, SuccessCode.UPDATE_WRITING_GOAL_SUCCESS);
     }
 
-    /*
-        TODO. [TR-YOO] 로그인 유저 검증 로직 필요
-    */
+
     @Operation(
             summary = "글쓰기 목표 삭제",
             description = "글쓰기 목표 삭제 API"
     )
     @DeleteMapping("")
     public ApiResponse<WritingGoalDeleteResponse> deleteWritingGoal(@Valid WritingGoalDeleteRequest request) {
+        jwtTokenProvider.checkUserOrAdmin();
 
         WritingGoalDeleteResponse response = writingGoalsService.removeWritingGoal(request);
